refactor tools into skill-based plugin architecture
Each tool (bash, edit, websearch) is now a self-contained skill under
skills/<name>/ with SKILL.md metadata and a scripts/ module. The engine
discovers and imports skills dynamically via importlib, dispatching with
a generic `skill.run(**args, **ctx)` call. Custom tools can be added by
dropping a new folder into skills/ — no engine changes needed.
