+++ pyproject.toml
@@ -18,3 +18,6 @@ build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
 packages = ["src/rlm"]
+
+[tool.hatch.build.targets.wheel.force-include]
+"skills" = "skills"
