+++ skills/bash/SKILL.md
@@ -1,0 +1,4 @@
+---
+name: bash
+description: Run a shell command and return its output. Use for file exploration, running tests, installing packages, and invoking `rlm` for sub-tasks.
+---
