+++ skills/bash/scripts/bash.py
@@ -1,0 +1,90 @@
+"""Bash tool — run shell commands."""
+
+from __future__ import annotations
+
+import os
+import re
+import subprocess
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from rlm.session import Session
+
+SCHEMA = {
+    "type": "function",
+    "function": {
+        "name": "bash",
+        "description": (
+            "Run a shell command and return its output. "
+            "Use for file exploration, running tests, installing packages, "
+            "and invoking `rlm` for sub-tasks."
+        ),
+        "parameters": {
+            "type": "object",
+            "properties": {
+                "command": {
+                    "type": "string",
+                    "description": "The shell command to execute.",
+                }
+            },
+            "required": ["command"],
+        },
+    },
+}
+
+_RLM_CMD_RE = re.compile(r"\brlm\s")
+
+
+def run(
+    command: str,
+    *,
+    cwd: str,
+    session: Session | None = None,
+    timeout: int = 120,
+    max_output: int = 8192,
+    **_,
+) -> str:
+    """Execute a bash command. Detects `rlm` invocations and sets up child sessions."""
+    env = os.environ.copy()
+
+    # Detect rlm sub-invocation → create child session dir
+    if session and _RLM_CMD_RE.search(command):
+        child_dir = session.child_dir()
+        env["RLM_SESSION_DIR"] = str(child_dir)
+        env["RLM_DEPTH"] = str(int(env.get("RLM_DEPTH", "0")) + 1)
+        # Propagate sub-tools if set
+        sub_tools = env.get("RLM_SUB_TOOLS")
+        if sub_tools:
+            env["RLM_TOOLS"] = sub_tools
+        session.log_sub_spawn(child_dir.name, command)
+
+    try:
+        result = subprocess.run(
+            command,
+            shell=True,
+            capture_output=True,
+            text=True,
+            timeout=timeout,
+            cwd=cwd,
+            env=env,
+        )
+    except subprocess.TimeoutExpired:
+        return f"[command timed out after {timeout}s]"
+
+    output = result.stdout
+    if result.stderr:
+        output += "\n" + result.stderr
+    if result.returncode != 0:
+        output += f"\n[exit code: {result.returncode}]"
+
+    # Truncate large output
+    if len(output) > max_output:
+        half = max_output // 2
+        total = len(output)
+        output = (
+            output[:half]
+            + f"\n... [output truncated, {total} chars total] ...\n"
+            + output[-half:]
+        )
+
+    return output
