+++ skills/edit/SKILL.md
@@ -1,0 +1,4 @@
+---
+name: edit
+description: Replace a unique string in a file. old_str must appear exactly once in the file.
+---
