+++ skills/websearch/SKILL.md
@@ -1,0 +1,4 @@
+---
+name: websearch
+description: Search Google via the Serper API. Accepts up to 10 queries in parallel. Returns titles, URLs, snippets, and knowledge-graph data.
+---
