+++ src/rlm/client.py
@@ -10,7 +10,11 @@ from rlm.types import TokenUsage
 def make_client() -> AsyncOpenAI:
     """Create an AsyncOpenAI client from environment variables."""
     base_url = os.environ.get("RLM_BASE_URL")
-    api_key = os.environ.get("RLM_API_KEY") or os.environ.get("OPENAI_API_KEY") or os.environ.get("ANTHROPIC_API_KEY")
+    api_key = (
+        os.environ.get("RLM_API_KEY")
+        or os.environ.get("OPENAI_API_KEY")
+        or os.environ.get("ANTHROPIC_API_KEY")
+    )
     return AsyncOpenAI(base_url=base_url, api_key=api_key)
